Define __unittest in scope of nose's assertion helpers to hide noisy stack frames. Closes #453.
I divided the tools into two buckets: trivial, which should not have their stack frames exposed in tracebacks; and nontrivial, which should. The tools module becomes a package, and everything stays backward-compatible.
