+++ AUTHORS
@@ -22,3 +22,4 @@ Takafumi Arakaki
 Peter Bengtsson
 Gary Donovan
 Brendan McCollam
+Erik Rose
