+++ nose/tools/nontrivial.py
@@ -1,34 +1,14 @@
-"""
-Tools for testing
------------------
+"""Tools not exempt from being descended into in tracebacks"""
 
-nose.tools provides a few convenience functions to make writing tests
-easier. You don't have to use them; nothing in the rest of nose depends
-on any of these methods.
-"""
-import re
 import time
-import unittest
 
 
-__all__ = ['ok_', 'eq_', 'make_decorator', 'raises', 'set_trace', 'timed',
-            'with_setup', 'TimeExpired', 'istest', 'nottest']
+__all__ = ['make_decorator', 'raises', 'set_trace', 'timed', 'with_setup',
+           'TimeExpired', 'istest', 'nottest']
 
 
 class TimeExpired(AssertionError):
     pass
-
-
-def ok_(expr, msg=None):
-    """Shorthand for assert. Saves 3 whole characters!
-    """
-    assert expr, msg
-
-
-def eq_(a, b, msg=None):
-    """Shorthand for 'assert a == b, "%r != %r" % (a, b)
-    """
-    assert a == b, msg or "%r != %r" % (a, b)
 
 
 def make_decorator(func):
@@ -168,27 +148,3 @@ def nottest(func):
     """
     func.__test__ = False
     return func
-
-#
-# Expose assert* from unittest.TestCase
-# - give them pep8 style names
-#
-caps = re.compile('([A-Z])')
-
-def pep8(name):
-    return caps.sub(lambda m: '_' + m.groups()[0].lower(), name)
-
-class Dummy(unittest.TestCase):
-    def nop():
-        pass
-_t = Dummy('nop')
-
-for at in [ at for at in dir(_t)
-            if at.startswith('assert') and not '_' in at ]:
-    pepd = pep8(at)
-    vars()[pepd] = getattr(_t, at)
-    __all__.append(pepd)
-
-del Dummy
-del _t
-del pep8
